usb: cdns3: make gadget code depend on kconfig symbol
The gadget code of the cdns3 core module will call into the gadget
module even if the gadget module is not enabled by
CONFIG_USB_CDNS3_GADGET.

FWIW, the code is already guarded by the CONFIG_DM_USB_GADGET symbol but
this might have been enabled for a different controller. I.e. this is
the case for the TI AM67A SoC which uses two different USB controllers
(the cadence one and the DWC3).

Add the correct "#if CONFIG_IS_ENABLED(USB_CDNS3_GADGET)."

Signed-off-by: Michael Walle <mwalle@kernel.org>