rng: Add OMAP TRNG driver
Add a driver for the True Random Number Generator (without DRBG support)
found on OMAP/K3 SoCs. The TRNG is part of the SA3UL crypto peripheral.
The TRNG DT node is also a subnode of the crypto node and this is why
this driver implements a dummy driver for the (yet) unsupported crypto
peripheral: to enumerate the child nodes. Otherwise the driver is really
straight forward.

This was tested on a AM67A SoC which contains a SA3UL, but it should
also work on SoCs with a SA2UL crypto peripheral. One would only need to
add the corresponding .compatibles to the driver.

Signed-off-by: Michael Walle <mwalle@kernel.org>