sa67: increase SPL size
Increase the R5 SPL size to include multiple DTBs. According to the TRM
the internal SRAM is 512kB but only 504kB are available for the loaded
image. Additionally, there is a file size restriction of 512kB for the
tiboot3.bin binary (because of the BootROM's FAT implementation). This
binary also includes the TIFS binary which is around 160kB at the
moment. Assuming a size of 192kB we're left with 504kB - 192kB = 312kB.

Signed-off-by: Michael Walle <mwalle@kernel.org>