sa67: select the correct device tree for SPL
Select the correct device tree for the SPL depending on the memory
strapping of the board.

We have to increase the heap because the device tree will be copied from
the FIT image to memory. Also we explicitly don't use compression
because that will require even more memory because it will extract the
whole FIT image into memory while the zipped image is still in memory!

Signed-off-by: Michael Walle <mwalle@kernel.org>