sa67: new environment loading logic
The failsafe as well as the production boot mode will need to have the
default environment, while the normal mode, will use an environment
stored in the MMC boot partitions. Until now, it was just possible to
erase (and save) the one in the MMC when operating from the failsafe
bootloader. Loading was cumbersome, as it has to be done by hand.
Introduce priorities for the failsafe boot mode. The highest prio has
the compiled-in default environment. Another environment driver, that is
the MMC one, can be selected by "env select MMC". If that is followed by
an "env load" the environment is then loaded from the MMC backend.

Signed-off-by: Michael Walle <mwalle@kernel.org>