sa67: Fixup DDR size when ECC is enabled
Port commit bc07851897bd ("board: ti: Pull redundant DDR functions to a
common location and Fixup DDR size when ECC is enabled") to sa67.

dram_init() and dram_init_banksize() cannot be removed since they
are only implemented in board/ti/common/k3-ddr.c, which we're not
including.

CONFIG_K3_INLINE_ECC depends on CONFIG_K3_DDRSS for its fixups. Thus we
can only have three different cases:

1. k3_ddrss = 0, inline_ecc = 0
2. k3_ddrss = 1, inline_ecc = 0
3. k3_ddrss = 1, inline_ecc = 1

Case (1) is true for the a53 builds (both u-boot spl and u-boot proper).
Until now, case (2) is used for the r5 build. With inline ECC support,
case (3) is also used in the r5 build as well as for the a53 u-boot.
Since we have split the modules into one for a53 u-boot and one for the
r5 loader, only case (1) and case (3) applies here. So the code should
be correct.

But that begs the question if fixup_memory_node() is needed in the r5
build. We're not sure, but it also isn't part of it right now...
and it seems to work just fine.

Signed-off-by: Lukas Zirpel <lukas@zirpel.de>