sysinfo: add driver for the Kontron SMARC-sAM67 board
The board has 3way strapping pins which are used to encode the board
variants. I.e. if it has a second ethernet interface or which RTC is
available.

Add a driver to query that information. The board code will use that
information to skip any unrelated device tree overlay. In the future,
this can also be used to select the correct base device tree.

Signed-off-by: Michael Walle <mwalle@kernel.org>